backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
boto3==1.33.13
botocore==1.33.13
certifi==2021.5.30
isort==4.3.21
jmespath==1.0.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.2
s3transfer==0.8.2
six==1.12.0
urllib3==1.26.18
wrapt==1.11.2
Flask
flask_restful
flask_cors
gunicorn
requests
python-dotenv